styling and fixed some bugs

+++ frontend/src/components/ProfilePage/Listing.js
@@ -9,14 +9,18 @@ const Listing = ({listing}) => {
   const goToReviews=(id)=>{
     history.push(`listing/${id}`)
   }
-  const deleteListing = (id)=> {
+
+  const deleteListingConfirmation =(id)=> {
+    let result = window.confirm('Are you sure you want to delete this listing?')
+    if(result){
     dispatch(deleteAListing(id))
+    }
   }
   return (
     <div className={`dashboard__listingCard`}>
     <p>{listing.description}</p>
    <button className='seeReviewsButton' onClick={()=>goToReviews(listing.id)}>See reviews</button>
-   <button className='deleteReviewsButton' onClick={()=> {deleteListing(listing.id)}}>Delete Listing</button>
+   <button className='deleteReviewsButton' onClick={()=> {deleteListingConfirmation(listing.id)}}>Delete Listing</button>
   </div>
   )
 }
